Easier config of motor control params and documentation

+++ shared/src/main/java/com/mcglynn/rvo/util/Constants.java
@@ -1,7 +1,7 @@
 package com.mcglynn.rvo.util;
 
 public class Constants {
-    public static final int MAX_THROTTLE = 100;
-    private static final int MAX_BRAKE = 100;
-    public static final int MAX_STEER = 1000;
+    public static final double MAX_THROTTLE = 100;
+    private static final double MAX_BRAKE = 100;
+    public static final double MAX_STEER = 1000;
 }
